Updated Parameters & Moved GSS to new file
Updated a lot of parameters so that ithe function now only takes a one dimensional array,
as well as updated parameters in a lot of other places.
This means replacing all "D" parameters with a specific dimension.

Commented out golden section search pending further thought on how to implement it given
restrictions on the typing of the array.